reorganizing qrec modules
